# perf: Faster symbol lookup