Two changes made:

1. Move getting symbols to XamlCodeGeneration instead of XamlFileGenerator. So we now have less calls.
2. Using `Lazy<T>` instead of `AsLockedMemoized`. Here is a benchmark for `AsLockedMemoized` vs `Lazy<T>`

```csharp
using System;
using System.Collections.Concurrent;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using Microsoft.CodeAnalysis;
using Microsoft.CodeAnalysis.CSharp;

public class Program
{
    public static void Main(string[] args)
    {
        _ = BenchmarkRunner.Run(typeof(Program).Assembly);
    }
}

public static class Funcs
{
    public static Func<TParam, TResult> Create<TParam, TResult>(Func<TParam, TResult> function) {
        return function;
    }
}

[MemoryDiagnoser]
public class Benchs
{
    private Func<string, ITypeSymbol> _findTypeByFullName;
    private Compilation _compilation;
    private Lazy<INamedTypeSymbol> _lazy;


    private ITypeSymbol SourceFindTypeByFullName(string fullName)
    {
        var symbol = _compilation.GetTypeByMetadataName(fullName);

        return symbol;
    }

    [GlobalSetup]
    public void Setup()
    {
        _compilation = CSharpCompilation.Create(null, new[] { SyntaxFactory.ParseSyntaxTree("namespace A.B.C { public class D { } }") });
        _findTypeByFullName = Funcs.Create<string, ITypeSymbol>(SourceFindTypeByFullName).AsLockedMemoized();
        _lazy = new Lazy<INamedTypeSymbol>(() => _compilation.GetTypeByMetadataName("A.B.C.D"));
    }

    [Benchmark(Baseline = true)]
    public void Baseline()
    {
        _findTypeByFullName("A.B.C.D");
    }

    [Benchmark]
    public void UsingLazy()
    {
        _ = _lazy.Value;
    }
}

public static class Ext
{
    public static Func<TKey, TResult> AsLockedMemoized<TKey, TResult>(this Func<TKey, TResult> func)
    {
        var values = new ConcurrentDictionary<TKey, TResult>();

        return (key) => values.GetOrAdd(key, func);
    }
}

```

Result on my machine:

|    Method |      Mean |     Error |    StdDev | Ratio | Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |----------:|----------:|----------:|------:|------:|------:|------:|----------:|
|  Baseline | 23.743 ns | 0.0698 ns | 0.0583 ns |  1.00 |     - |     - |     - |         - |
| UsingLazy |  3.529 ns | 0.0187 ns | 0.0156 ns |  0.15 |     - |     - |     - |         - |

![image](https://user-images.githubusercontent.com/31348972/221410676-97ca0114-0ae3-40d9-a258-9ca834bc5f8c.png)