@Youssef1313 2023-02-26
@jeromelaban @spouliot I was expecting an improvement, but I'm surprised it is *so large* that I'm suspecting I might be benchmarking incorrectly? Can you review the benchmark please?

My benchmark doesn't put into account that some symbols might not be accessed, which will actually make `Lazy<T>` approach even MORE FASTER!