@Youssef1313 2023-02-27
> I'm not familiar with `AsLockedMemoized` and will have to look it up...

Basically ConcurrentDictionary under the hoods