@Youssef1313 2023-02-27
https://github.com/unoplatform/uno/blob/1a45d03fbefd938a0e880edcca99d895cf651586/src/Uno.Foundation/Uno.Core.Extensions/Uno.Core.Extensions/FuncMemoizeExtensions.cs#L231-L267

The relevant conditional code of the above snippet is that:

```csharp
 			var values = new ConcurrentDictionary<TKey, TResult>(); 
  
 			return (key) => values.GetOrAdd(key, func); 
```