@jeromelaban 2023-02-27
The original code from this part did not have Lazy available (as far as I remember), and if can use Lazy now, way to go :)